Move php code to filter.php
