+++ share.php
@@ -1,41 +1,6 @@
 <?php
     include "layout/header.php";
 
-    function reArrayFiles(&$file_post) {
-
-        $file_ary = array();
-        $file_count = count($file_post['name']);
-        $file_keys = array_keys($file_post);
-    
-        for ($i=0; $i<$file_count; $i++) {
-            foreach ($file_keys as $key) {
-                $file_ary[$i][$key] = $file_post[$key][$i];
-            }
-        }
-    
-        return $file_ary;
-    }
-    
-    
-    if($_SERVER['REQUEST_METHOD'] == 'POST') {
-        $target_dir = "uploads/";
-    
-        if (!file_exists($target_dir)) {
-            mkdir('uploads/', 0777, true);
-        }
-
-        if ($_FILES['images']) {
-                $file_ary = reArrayFiles($_FILES['images']);
-            
-            foreach ($file_ary as $file) {
-                $target_file = $target_dir . basename($file['name']);
-                move_uploaded_file($file['tmp_name'], $target_file);
-            }
-                
-        }
-    }
-    
-
     $directory = "uploads/";
     $scanned_directory = array_diff(scandir($directory), array('..', '.'));
 ?>
